@@ -9,7 +9,7 @@ author: Lorenzo Pongetti <lorenzo.pongetti@gmail.com>
 homepage: https://github.com/lpongetti/bubble_chart
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
