@@ -1,7 +1,7 @@
 class BubbleNodeBase {
-  double radius;
-  double x = 0;
-  double y = 0;
+  double? radius;
+  double? x = 0;
+  double? y = 0;
 
   BubbleNodeBase({this.radius, this.x, this.y});
 }
